# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: GmbH API
  version: 1.2.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/areas:
    get:
      tags:
        - Area
      summary: Request all areas
      responses:
        '200':
          description: Areas list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Area
      summary: Create one area
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AreaCreateRequest'
      responses:
        '200':
          description: Area
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/areas/{id}:
    get:
      tags:
        - Area
      summary: Request area by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Area
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaResponse'
    put:
      tags:
        - Area
      summary: Update area
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AreaUpdateRequest'
      responses:
        '200':
          description: Area
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Area
      summary: Delete area
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Empty response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/categories:
    get:
      tags:
        - Category
      summary: Request all categories
      responses:
        '200':
          description: Categories list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Category
      summary: Create one category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '200':
          description: Category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/categories/{id}:
    get:
      tags:
        - Category
      summary: Request category by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Category
      summary: Update category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateRequest'
      responses:
        '200':
          description: Category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Category
      summary: Delete category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Empty response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-access-token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        errors:
          type: object
          properties:
            msg:
              type: string
          required:
            - msg
      required:
        - errors
...
    Area:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        short:
          type: string
        enabled:
          type: boolean
        tables:
          type: array
          items:
            type: string
        users:
          type: array
          items:
            type: string
    AreaResponse:
      type: object
      properties:
        area:
          $ref: '#/components/schemas/Area'
      required:
        - area
    AreaListResponse:
      type: object
      properties:
        areas:
          type: array
          items:
            $ref: '#/components/schemas/Area'
      required:
        - areas
    AreaCreateRequest:
      type: object
      properties:
        area:
          $ref: '#/components/schemas/Area'
      required:
        - area
    AreaUpdateRequest:
      type: object
      properties:
        area:
          $ref: '#/components/schemas/Area'
      required:
        - area
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        description:
          type: string
        icon:
          type: string
          description: Icon name for category display
        color:
          type: string
          description: Hex color for category display
          example: '#94a3b8'
        showAmount:
          type: boolean
        printer:
          type: string
        categories:
          type: array
          items:
            type: string
        category:
          type: string
    CategoryResponse:
      type: object
      properties:
        category:
          $ref: '#/components/schemas/Category'
      required:
        - category
    CategoryListResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
      required:
        - categories
    CategoryCreateRequest:
      type: object
      properties:
        category:
          $ref: '#/components/schemas/Category'
      required:
        - category
    CategoryUpdateRequest:
      type: object
      properties:
        category:
          $ref: '#/components/schemas/Category'
      required:
        - category
...